various fixes for notes in OZ audit (#1437)
* add names to all returns for clarity

* fix some docstrings (reported by oz)

* fix wrong or unused dependencies

identified by oz audit

* consistent use of implicit `int` and `uint` types

everything is uint256 or int256 for max clarity now

recocomended by oz audit

* remove unnecessary downcast

* improve revert error consistencies

* fix compile errro

* fix variable shadowing

* fix wrong function visibility

`CollateralModule` `deposit` and `withdraw` should both be external

* fix incorrect naming of return parameter

* lint fix

+++ protocol/synthetix/contracts/modules/core/MarketManagerModule.sol
@@ -159,7 +159,7 @@ contract MarketManagerModule is IMarketManagerModule {
      */
     function distributeDebtToPools(
         uint128 marketId,
-        uint maxIter
+        uint256 maxIter
     ) external override returns (bool) {
         return Market.load(marketId).distributeDebtToPools(maxIter);
     }
