+++ protocol/synthetix/contracts/storage/Account.sol
@@ -38,7 +38,11 @@ library Account {
     /**
      * @dev Thrown when the requested operation requires an activity timeout before the
      */
-    error AccountActivityTimeoutPending(uint128 accountId, uint currentTime, uint requiredTime);
+    error AccountActivityTimeoutPending(
+        uint128 accountId,
+        uint256 currentTime,
+        uint256 requiredTime
+    );
 
     struct Data {
         /**
@@ -94,7 +98,7 @@ library Account {
     }
 
     /**
-     * @dev Returns information about the total collateral assigned, deposited, and locked by the account, and the given collateral type.
+     * @dev Given a collateral type, returns information about the total collateral assigned, deposited, and locked by the account
      */
     function getCollateralTotals(
         Data storage self,
@@ -176,7 +180,7 @@ library Account {
     function loadAccountAndValidatePermissionAndTimeout(
         uint128 accountId,
         bytes32 permission,
-        uint timeout
+        uint256 timeout
     ) internal view returns (Data storage account) {
         account = Account.load(accountId);
 
@@ -184,7 +188,7 @@ library Account {
             revert PermissionDenied(accountId, permission, msg.sender);
         }
 
-        uint endWaitingPeriod = account.lastInteraction + timeout;
+        uint256 endWaitingPeriod = account.lastInteraction + timeout;
         if (block.timestamp < endWaitingPeriod) {
             revert AccountActivityTimeoutPending(accountId, block.timestamp, endWaitingPeriod);
         }
